Introduce new VGA driver
