+++ kernel/src/idt.c
@@ -4,6 +4,7 @@
  */
 
 #include <idt.h>
+#include <vga.h>
 
 IdtEntry idt[256];
 IdtPointer idtPointer;
@@ -25,4 +26,6 @@ void InstallIdt(void)
 	idtPointer.base = (uint32_t)&idt;
 
 	IdtLoad();
+
+	PutStr("Installed IDT\n");
 }
