+++ kernel/src/include/isr.h
@@ -1,0 +1,6 @@
+#ifndef ISR_H
+#define ISR_H
+
+void InstallIsr(void);
+
+#endif
