+++ kernel/src/isr.c
@@ -1,0 +1,12 @@
+/*
+ * isr.c - Interrupt Service Routines
+ * Author(s) - amrix
+ */
+
+#include <isr.h>
+#include <vga.h>
+
+void InstallIsr(void)
+{
+	PutStr("Installed ISRs\n");
+}
