+++ kernel/src/main.c
@@ -5,13 +5,17 @@
 
 #include <vga.h>
 #include <idt.h>
+#include <isr.h>
 
 void CEntry(void) 
 {
-	EnableCursor();
+	InitVga();
+	
+EnableCursor();
 	PutStr("ZecelOS Kernel\n\n");
 
 	InstallIdt();
+	InstallIsr();
 
 	while (1)
 		asm volatile("hlt");
